# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Invest Tracker API
  version: 1.0.0
  description: Phase 1 portfolio, account and instrument API. Authentication is deferred; deployment must not expose these endpoints without an appropriate authentication boundary.
servers:
  - url: /api/v1
paths:
...
  /portfolios/{portfolioId}/accounts/{accountId}:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
    get:
      operationId: getAccount
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: archiveAccount
      responses:
        '204': { description: Archived }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/accounts/{accountId}/positions:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
    get:
      operationId: listPositions
      responses:
        '200':
          description: Account active positions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createPosition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/accounts/{accountId}/positions/{positionId}:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
      - { $ref: '#/components/parameters/PositionId' }
    get:
      operationId: getPosition
      responses:
        '200':
          description: Position details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      operationId: updatePosition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionUpdate' }
      responses:
        '200':
          description: Updated position
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: archivePosition
      responses:
        '204': { description: Archived }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/accounts/{accountId}/positions/{positionId}/lots:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
      - { $ref: '#/components/parameters/PositionId' }
    get:
      operationId: getPositionLots
      responses:
        '200':
          description: Position tax lots and cost basis details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PositionLotsDetail' }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/accounts/{accountId}/positions/{positionId}/performance:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
      - { $ref: '#/components/parameters/PositionId' }
    get:
      operationId: getPositionPerformance
      responses:
        '200':
          description: Position performance details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PositionPerformance' }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/accounts/{accountId}/positions/performance:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
      - { $ref: '#/components/parameters/AccountId' }
    get:
      operationId: listAccountPositionsPerformance
      parameters:
        - in: query
          name: includeClosed
          required: false
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Account positions performance
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PositionPerformance' }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/positions/performance:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
    get:
      operationId: listPortfolioPositionsPerformance
      parameters:
        - in: query
          name: includeClosed
          required: false
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Portfolio positions performance
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PositionPerformance' }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/positions:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
    get:
      operationId: listPortfolioPositions
      responses:
        '200':
          description: Portfolio active positions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/Problem' }
  /portfolios/{portfolioId}/positions/recalculate:
    parameters:
      - { $ref: '#/components/parameters/PortfolioId' }
    post:
      operationId: recalculatePortfolioPositions
      responses:
        '200':
          description: Portfolio positions recalculation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PositionRecalculateResponse' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /instruments:
    get:
      operationId: listInstruments
      responses:
        '200':
          description: Instruments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Instrument' }
    post:
      operationId: createInstrument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InstrumentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Instrument' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /instruments/{instrumentId}/refresh-price:
    parameters:
      - { $ref: '#/components/parameters/InstrumentId' }
    post:
      operationId: refreshInstrumentPrice
      responses:
        '200':
          description: Refreshed instrument with latest quote
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Instrument' }
        '404': { $ref: '#/components/responses/Problem' }
  /instruments/{instrumentId}:
    parameters:
      - { $ref: '#/components/parameters/InstrumentId' }
    get:
      operationId: getInstrument
      responses:
        '200':
          description: Instrument details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Instrument' }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      operationId: updateInstrument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InstrumentCreate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Instrument' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    PortfolioId:
      in: path
      name: portfolioId
      required: true
      schema: { type: string, format: uuid }
    AccountId:
      in: path
      name: accountId
      required: true
      schema: { type: string, format: uuid }
    InstrumentId:
      in: path
      name: instrumentId
      required: true
      schema: { type: string, format: uuid }
    PositionId:
      in: path
      name: positionId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Account:
      type: object
      required: [id, portfolioId, name, brokerName, accountCurrency, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        portfolioId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        brokerName: { type: string, minLength: 1, maxLength: 120 }
        accountCurrency: { type: string, pattern: '^[A-Z]{3}$' }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AccountCreate:
      type: object
      required: [name, brokerName, accountCurrency]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        brokerName: { type: string, minLength: 1, maxLength: 120 }
        accountCurrency: { type: string, pattern: '^[A-Za-z]{3}$' }
    Instrument:
      type: object
      required: [id, name, assetClass, currency, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 160 }
        assetClass: { $ref: '#/components/schemas/AssetClass' }
        ticker: { type: string, maxLength: 32, nullable: true }
        isin: { type: string, maxLength: 12, nullable: true }
        exchange: { type: string, maxLength: 80, nullable: true }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        latestPrice: { type: number, nullable: true }
        priceCurrency: { type: string, pattern: '^[A-Za-z]{3}$', nullable: true }
        priceAsOf: { type: string, format: date-time, nullable: true }
        isStale: { type: boolean, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InstrumentCreate:
      type: object
      required: [name, assetClass, currency]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        assetClass: { $ref: '#/components/schemas/AssetClass' }
        ticker: { type: string, maxLength: 32, nullable: true }
        isin: { type: string, maxLength: 12, nullable: true }
        exchange: { type: string, maxLength: 80, nullable: true }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
    Position:
      type: object
      required: [id, accountId, instrumentId, instrumentName, assetClass, quantity, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        instrumentName: { type: string }
        instrumentTicker: { type: string, nullable: true }
        instrumentIsin: { type: string, nullable: true }
        assetClass: { $ref: '#/components/schemas/AssetClass' }
        quantity: { type: number, minimum: 0 }
        costBasisAmount: { type: number, minimum: 0, nullable: true }
        costBasisCurrency: { type: string, pattern: '^[A-Z]{3}$', nullable: true }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PositionCreate:
      type: object
      required: [instrumentId, quantity]
      properties:
        instrumentId: { type: string, format: uuid }
        quantity: { type: number, minimum: 0 }
        costBasisAmount: { type: number, minimum: 0, nullable: true }
        costBasisCurrency: { type: string, pattern: '^[A-Za-z]{3}$', nullable: true }
    PositionUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: number, minimum: 0 }
        costBasisAmount: { type: number, minimum: 0, nullable: true }
        costBasisCurrency: { type: string, pattern: '^[A-Za-z]{3}$', nullable: true }
    PositionLot:
      type: object
      required: [lotId, acquisitionDate, originalQuantity, remainingQuantity, unitCostAmount, totalCostAmount, currency]
      properties:
        lotId: { type: string, format: uuid }
        transactionId: { type: string, format: uuid, nullable: true }
        acquisitionDate: { type: string, format: date-time }
        originalQuantity: { type: number, minimum: 0 }
        remainingQuantity: { type: number, minimum: 0 }
        unitCostAmount: { type: number }
        totalCostAmount: { type: number }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
    LotDisposal:
      type: object
      required: [transactionId, lotId, disposalDate, quantity, costBasis, proceeds, realizedGainLoss, currency]
      properties:
        transactionId: { type: string, format: uuid }
        lotId: { type: string, format: uuid }
        disposalDate: { type: string, format: date-time }
        quantity: { type: number }
        costBasis: { type: number }
        proceeds: { type: number }
        realizedGainLoss: { type: number }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
    PositionLotsDetail:
      type: object
      required: [positionId, accountId, instrumentId, costBasisMethod, totalQuantity, totalCostBasisAmount, currency, averageUnitCostAmount, realizedGainLossAmount, openLots]
      properties:
        positionId: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid }
        costBasisMethod: { type: string, enum: [FIFO, LIFO, AVERAGE_COST] }
        totalQuantity: { type: number }
        totalCostBasisAmount: { type: number }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
        averageUnitCostAmount: { type: number }
        realizedGainLossAmount: { type: number }
        openLots:
          type: array
          items: { $ref: '#/components/schemas/PositionLot' }
    PositionPerformance:
      type: object
      required: [positionId, accountId, accountName, instrumentId, instrumentName, assetClass, status, currentQuantity, totalBoughtQuantity, totalSoldQuantity, totalInvestedAmount, totalProceedsAmount, netTotalReturnAmount, totalReturnPercentage, currency]
      properties:
        positionId: { type: string, format: uuid, nullable: true }
        accountId: { type: string, format: uuid }
        accountName: { type: string }
        instrumentId: { type: string, format: uuid }
        instrumentName: { type: string }
        ticker: { type: string, nullable: true }
        isin: { type: string, nullable: true }
        assetClass: { $ref: '#/components/schemas/AssetClass' }
        status: { type: string, enum: [ACTIVE, CLOSED, ARCHIVED] }
        currentQuantity: { type: number }
        totalBoughtQuantity: { type: number }
        totalSoldQuantity: { type: number }
        averageBuyPrice: { type: number }
        averageSellPrice: { type: number }
        totalInvestedAmount: { type: number }
        totalProceedsAmount: { type: number }
        currentCostBasis: { type: number }
        currentPrice: { type: number }
        currentMarketValue: { type: number }
        realizedGainLoss: { type: number }
        unrealizedGainLoss: { type: number }
        dividendIncome: { type: number }
        fees: { type: number }
        taxes: { type: number }
        netTotalReturnAmount: { type: number }
        totalReturnPercentage: { type: number }
        currency: { type: string, pattern: '^[A-Za-z]{3}$' }
        nativePrice: { type: number, nullable: true }
        nativeCurrency: { type: string, nullable: true }
        nativeNetTotalReturnAmount: { type: number, nullable: true }
        nativeReturnPercentage: { type: number, nullable: true }
        openLots:
          type: array
          items: { $ref: '#/components/schemas/PositionLot' }
        disposals:
          type: array
          items: { $ref: '#/components/schemas/LotDisposal' }
        transactions:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
    PositionRecalculateResponse:
      type: object
      required: [portfolioId, recalculatedPositionsCount, message]
      properties:
        portfolioId: { type: string, format: uuid }
        recalculatedPositionsCount: { type: integer }
        message: { type: string }
    TransactionType:
      type: string
      enum: [BUY, SELL, DIVIDEND, FEE, DEPOSIT, WITHDRAWAL, INTEREST, STOCK_SPLIT, REVERSE_STOCK_SPLIT, TRANSFER]
    Transaction:
      type: object
      required: [id, accountId, type, tradeDate, grossAmount, netAmount, currency, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        instrumentId: { type: string, format: uuid, nullable: true }
        instrumentName: { type: string, nullable: true }
        instrumentTicker: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/TransactionType' }
        tradeDate: { type: string, format: date-time }
        settlementDate: { type: string, format: date-time, nullable: true }
        quantity: { type: number, nullable: true }
        price: { type: number, nullable: true }
        grossAmount: { type: number }
        feeAmount: { type: number, nullable: true }
        taxAmount: { type: number, nullable: true }
        netAmount: { type: number }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        fxRate: { type: number, nullable: true }
        counterCurrency: { type: string, pattern: '^[A-Z]{3}$', nullable: true }
        notes: { type: string, maxLength: 255, nullable: true }
        status: { type: string, enum: [COMPLETED, CORRECTED] }
        correctionOfTransactionId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AssetClass:
      type: string
      enum: [STOCK, ETF, MUTUAL_FUND, BOND, REIT, CRYPTO, CASH, OTHER]
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errors:
          type: array
          items: { type: string }
...